cleaned up installation and vimrc

+++ install.sh
@@ -5,41 +5,59 @@ echo "Setting up your Mac..."
 dir=~/dotfiles                    # dotfiles directory
 olddir=~/dotfiles_old             # old dotfiles backup directory
 files="zshrc vimrc"               # list of files/folders to symlink in homedir
-numsteps=5
+numsteps=6
+currentstep=1
 
 ##### Install Xcode cli
-echo "Step 1/$numsteps: Install Xcode"
-xcode-select --install
+echo "Step $currentstep/$numsteps: Install Xcode"
+((currentstep++))
 
-##### Check for Homebrew and install it if it is not found
-echo "Step 2/$numsteps: Install Homebrew and dependencies listed in the Brewfile"
+if test $(xcode-select -p); then
+    echo 'Xcode is already installed'
+else
+    xcode-select --install
+fi
+
+##### Check for Homebrew and install it if not found
+echo "Step $currentstep/$numsteps: Install Homebrew and dependencies listed in the Brewfile"
+((currentstep++))
+
 if test ! $(which brew); then
   /usr/bin/ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
 fi
 
 # Update Homebrew recipes
 brew update
-brew upgrade
 
 # Install all our dependencies with bundle (See Brewfile)
 brew tap homebrew/bundle
+brew install https://raw.githubusercontent.com/Homebrew/homebrew-core/f2a764ef944b1080be64bd88dca9a1d80130c558/Formula/python.rb
 brew bundle
 
 ##### Install Python dependencies
-echo "Step 3/$numsteps: Install python dependencies"
+echo "Step $currentstep/$numsteps: Install python dependencies"
+((currentstep++))
+
 pip3 install -r $dir/requirements.txt
 
 #### Install Powerlevel9k & fonts
-echo "Step 4/$numsteps: Install Powerlevel9k and powerline fonts"
-git clone https://github.com/bhilburn/powerlevel9k.git ~/powerlevel9k
-git clone https://github.com/powerline/fonts.git --depth=1
-cd fonts
-./install.sh
-cd ..
-rm -rf fonts
+echo "Step $currentstep/$numsteps: Install Powerlevel9k and powerline fonts"
+((currentstep++))
+
+if [ -d ~/powerlevel9k ]; then
+    echo 'Powerlevel9k is already installed'
+else
+    git clone https://github.com/bhilburn/powerlevel9k.git ~/powerlevel9k
+    git clone https://github.com/powerline/fonts.git --depth=1
+    cd fonts
+    ./install.sh
+    cd ..
+    rm -rf fonts
+fi
 
 ##### Create symlinks
-echo "Step 5/$numsteps: Create symlinks to the repo's dotfiles"
+echo "Step $currentstep/$numsteps: Create symlinks to the repo's dotfiles"
+((currentstep++))
 
 # create dotfiles_old in homedir
 echo "Creating $olddir for backup of any existing dotfiles in ~"
@@ -58,7 +76,18 @@ done
 echo "Create symlink for kitty config file"
 ln -Fs $dir/kitty.conf ~/.config/kitty/kitty.conf
 
+##### Configure vim
+mkdir -p ~/.vim/tmp # backup files
+mkdir ~/.vim/bundle # plugins
+
 ##### Set zsh as default shell
-echo "Step 6/$numsteps: Set zsh as default shell"
-echo "$(which zsh)" | sudo tee -a /etc/shells
-chsh -s $(which zsh)
+echo "Step $currentstep/$numsteps: Setting zsh as default shell"
+((currentstep++))
+
+if [ $SHELL = $(which zsh) ]; then
+    echo 'zsh is already the default shell!'
+else
+    echo "$(which zsh)" | sudo tee -a /etc/shells
+    chsh -s $(which zsh)
+fi
+
